+++ kinappserver/__init__.py
@@ -26,7 +26,7 @@ db = SQLAlchemy(app)
 
 # TODO remove this on production
 admin = Admin(app, name='KinApp', template_mode='bootstrap3')
-from kinappserver.model import User, UserAppData, UserTaskResults, Task, Transaction
+from kinappserver.models import User, UserAppData, UserTaskResults, Task, Transaction
 admin.add_view(ModelView(User, db.session))
 admin.add_view(ModelView(UserAppData, db.session))
 admin.add_view(ModelView(UserTaskResults, db.session))
