+++ kinappserver/models/__init__.py
@@ -1,0 +1,3 @@
+from .transaction import *
+from .user import *
+from .task import *
