+++ kinappserver/tests/task.py
@@ -13,7 +13,7 @@ import testing.postgresql
 from flask import Flask
 
 import kinappserver
-from kinappserver import db, config, model
+from kinappserver import db, config, models
 
 USER_ID_HEADER = "X-USERID"
 
@@ -72,7 +72,7 @@ class Tester(unittest.TestCase):
                             content_type='application/json')
         self.assertEqual(resp.status_code, 200)
 
-        print(model.list_all_task_data())
+        print(models.list_all_task_data())
 
         # register a user
         userid = uuid.uuid4()
