+++ setup.py
@@ -1,8 +1,8 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
     name='kinappserver',
-    packages=['kinappserver'],
+    packages=find_packages(),
     include_package_data=True,
     install_requires=[
         'flask'
